+++ diffsims/tests/crystallography/test_diffracting_vector.py
@@ -65,3 +65,17 @@ class TestDiffractingVector:
         r, t = dv.to_flat_polar()
         assert np.allclose(r, [np.sqrt(2), np.sqrt(2), 0.70710678, 0.70710678])
         assert np.allclose(t, [np.pi / 4, np.pi / 4, -np.pi / 4, -np.pi / 4])
+
+    def test_get_lattice_basis_rotation(self, ferrite_phase):
+        """Rotation matrix to align the lattice basis with the Cartesian
+        basis is correct.
+        """
+        rlv = DiffractingVector(ferrite_phase, hkl=[[1, 1, 1], [2, 0, 0]])
+        rot = rlv.basis_rotation
+        assert np.allclose(rot.to_matrix(), np.eye(3))
+
+    def test_rotation_with_basis_raises(self, ferrite_phase):
+        rlv = DiffractingVector(ferrite_phase, hkl=[[1, 1, 1], [2, 0, 0]])
+        rot = Rotation.from_euler([[90, 90, 0], [90, 90, 1]], degrees=True)
+        with pytest.raises(ValueError):
+            rlv.rotate_with_basis(rotation=rot)
